+++ src/IoTSharp.Data.Taos/Protocols/TDWebSocket/WSSchemalessRsp.cs
@@ -2,8 +2,8 @@
 {
     public class WSSchemalessRsp: WSActionRsp
     {
-        public int req_id { get; set; }
-        public int timing { get; set; }
+        public long req_id { get; set; }
+        public long timing { get; set; }
 
     }
 }
